+++ src-opam/dockerfile_distro.ml
@@ -19,10 +19,10 @@
 open Astring
 
 type t = [
-  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `Latest ]
+  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12 | `Latest ]
   | `CentOS of [ `V6 | `V7 | `V8 | `Latest ]
   | `Debian of [ `V10 | `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
-  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `Latest ]
+  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `V32 | `Latest ]
   | `OracleLinux of [ `V7 | `Latest ]
   | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `Latest ]
   | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `V19_04 | `V19_10 | `V20_04 | `LTS | `Latest ]
@@ -35,11 +35,11 @@ type status = [
 ] [@@deriving sexp]
 
 let distros = [
-  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `V3_10; `Alpine `V3_11; `Alpine `Latest;
+  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `V3_10; `Alpine `V3_11; `Alpine `V3_12; `Alpine `Latest;
   `CentOS `V6; `CentOS `V7; `CentOS `V8; `CentOS `Latest;
   `Debian `V10; `Debian `V9; `Debian `V8; `Debian `V7;
   `Debian `Stable; `Debian `Testing; `Debian `Unstable;
-  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `V30; `Fedora `V31; `Fedora `Latest;
+  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `V30; `Fedora `V31; `Fedora `V32; `Fedora `Latest;
   `OracleLinux `V7; `OracleLinux `Latest;
   `OpenSUSE `V42_1; `OpenSUSE `V42_2; `OpenSUSE `V42_3; `OpenSUSE `V15_0; `OpenSUSE `V15_1; `OpenSUSE `Latest;
   `Ubuntu `V12_04; `Ubuntu `V14_04; `Ubuntu `V15_04; `Ubuntu `V15_10;
@@ -47,9 +47,9 @@ let distros = [
   `Ubuntu `Latest; `Ubuntu `LTS ]
 
 let distro_status (d:t) : status = match d with
-  | `Alpine ( `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9) -> `Deprecated
-  | `Alpine `V3_10 -> `Active `Tier2
-  | `Alpine `V3_11 -> `Active `Tier1
+  | `Alpine ( `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10) -> `Deprecated
+  | `Alpine `V3_11 -> `Active `Tier2
+  | `Alpine `V3_12 -> `Active `Tier1
   | `Alpine `Latest -> `Alias (`Alpine `V3_11)
   | `CentOS (`V7 | `V8) -> `Active `Tier2
   | `CentOS `V6 -> `Deprecated
@@ -63,7 +63,8 @@ let distro_status (d:t) : status = match d with
   | `Debian `Unstable -> `Active `Tier2
   | `Fedora ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30) -> `Deprecated
   | `Fedora `V31 -> `Active `Tier2
-  | `Fedora `Latest -> `Alias (`Fedora `V31)
+  | `Fedora `V32 -> `Active `Tier2
+  | `Fedora `Latest -> `Alias (`Fedora `V31) (* TODO until opam-repo fixed for older ocaml *)
   | `OracleLinux `V7 -> `Active `Tier2
   | `OracleLinux `Latest -> `Alias (`OracleLinux `V7)
   | `OpenSUSE (`V42_1 | `V42_2 | `V42_3 | `V15_0) -> `Deprecated
@@ -90,8 +91,8 @@ module OV = Ocaml_version
 
 let distro_arches ov (d:t) =
   match resolve_alias d, ov with
-  | `Debian (`V9|`V10), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64; `Ppc64le; `Aarch32 ]
-  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
+  | `Debian (`V9|`V10), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `I386; `X86_64; `Aarch64; `Ppc64le; `Aarch32 ]
+  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `I386; `X86_64; `Aarch64 ]
   | `Ubuntu (`V16_04|`V18_04|`V17_10|`V18_10|`V19_04|`V19_10|`V20_04), ov when OV.(compare Releases.v4_05_0 ov) = -1  -> [ `X86_64; `Aarch64 ]
   | _ -> [ `X86_64 ]
 
@@ -140,6 +141,7 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Alpine `V3_9 -> Some "4.06.1"
   |`Alpine `V3_10 -> Some "4.07.0"
   |`Alpine `V3_11 -> Some "4.08.1"
+  |`Alpine `V3_12 -> Some "4.08.1"
   |`Fedora `V21 -> Some "4.01.0"
   |`Fedora `V22 -> Some "4.02.0"
   |`Fedora `V23 -> Some "4.02.2"
@@ -151,6 +153,7 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Fedora `V29 -> Some "4.07.0"
   |`Fedora `V30 -> Some "4.07.0"
   |`Fedora `V31 -> Some "4.08.1"
+  |`Fedora `V32 -> Some "4.10.0"
   |`CentOS `V6 -> Some "3.11.2"
   |`CentOS `V7 -> Some "4.01.0"
   |`CentOS `V8 -> Some "4.07.0"
@@ -204,6 +207,7 @@ let tag_of_distro (d:t) = match d with
   |`Fedora `V29 -> "fedora-29"
   |`Fedora `V30 -> "fedora-30"
   |`Fedora `V31 -> "fedora-31"
+  |`Fedora `V32 -> "fedora-32"
   |`OracleLinux `V7 -> "oraclelinux-7"
   |`OracleLinux `Latest -> "oraclelinux"
   |`Alpine `V3_3 -> "alpine-3.3"
@@ -215,6 +219,7 @@ let tag_of_distro (d:t) = match d with
   |`Alpine `V3_9 -> "alpine-3.9"
   |`Alpine `V3_10 -> "alpine-3.10"
   |`Alpine `V3_11 -> "alpine-3.11"
+  |`Alpine `V3_12 -> "alpine-3.12"
   |`Alpine `Latest -> "alpine"
   |`OpenSUSE `V42_1 -> "opensuse-42.1"
   |`OpenSUSE `V42_2 -> "opensuse-42.2"
@@ -260,6 +265,7 @@ let distro_of_tag x : t option = match x with
   |"fedora-29" -> Some (`Fedora `V29)
   |"fedora-30" -> Some (`Fedora `V30)
   |"fedora-31" -> Some (`Fedora `V31)
+  |"fedora-32" -> Some (`Fedora `V32)
   |"fedora" -> Some (`Fedora `Latest)
   |"oraclelinux-7" -> Some (`OracleLinux `V7)
   |"oraclelinux" -> Some (`OracleLinux `Latest)
@@ -272,6 +278,7 @@ let distro_of_tag x : t option = match x with
   |"alpine-3.9" -> Some (`Alpine `V3_9)
   |"alpine-3.10" -> Some (`Alpine `V3_10)
   |"alpine-3.11" -> Some (`Alpine `V3_11)
+  |"alpine-3.12" -> Some (`Alpine `V3_12)
   |"alpine" -> Some (`Alpine `Latest)
   |"opensuse-42.1" -> Some (`OpenSUSE `V42_1)
   |"opensuse-42.2" -> Some (`OpenSUSE `V42_2)
@@ -318,6 +325,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Fedora `V29 -> "Fedora 29"
   |`Fedora `V30 -> "Fedora 30"
   |`Fedora `V31 -> "Fedora 31"
+  |`Fedora `V32 -> "Fedora 32"
   |`OracleLinux `V7 -> "OracleLinux 7"
   |`Alpine `V3_3 -> "Alpine 3.3"
   |`Alpine `V3_4 -> "Alpine 3.4"
@@ -328,6 +336,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Alpine `V3_9 -> "Alpine 3.9"
   |`Alpine `V3_10 -> "Alpine 3.10"
   |`Alpine `V3_11 -> "Alpine 3.11"
+  |`Alpine `V3_12 -> "Alpine 3.12"
   |`OpenSUSE `V42_1 -> "OpenSUSE 42.1"
   |`OpenSUSE `V42_2 -> "OpenSUSE 42.2"
   |`OpenSUSE `V42_3 -> "OpenSUSE 42.3"
@@ -369,9 +378,9 @@ let package_manager (t:t) =
   |`Alpine _ -> `Apk
   |`OpenSUSE _ -> `Zypper
 
-let base_distro_tag d =
+let base_distro_tag ?(arch=`X86_64) d =
   match resolve_alias d with
-  | `Alpine v ->
+  | `Alpine v -> begin
         let tag =
           match v with
           | `V3_3 -> "3.3"
@@ -383,10 +392,14 @@ let base_distro_tag d =
           | `V3_9 -> "3.9"
           | `V3_10 -> "3.10"
           | `V3_11 -> "3.11"
+          | `V3_12 -> "3.12"
           | `Latest -> assert false
         in
-        "alpine", tag
-   | `Debian v ->
+        match arch with
+        | `I386 -> "i386/alpine", tag
+        | _ -> "alpine", tag
+   end
+   | `Debian v -> begin
         let tag =
           match v with
           | `V7 -> "7"
@@ -397,7 +410,10 @@ let base_distro_tag d =
           | `Unstable -> "unstable"
           | `Stable -> assert false
         in
-        "debian",tag
+        match arch with
+        | `I386 -> "i386/debian", tag
+        | _ -> "debian", tag
+    end
     | `Ubuntu v ->
         let tag =
           match v with
@@ -434,6 +450,7 @@ let base_distro_tag d =
           | `V29 -> "29"
           | `V30 -> "30"
           | `V31 -> "31"
+          | `V32 -> "32"
           | `Latest -> assert false
         in
         "fedora", tag