+++ src-opam/dockerfile_distro.mli
@@ -23,10 +23,10 @@
 (** {2 Known distributions and OCaml variants} *)
 
 type t = [
-  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `Latest ]
+  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12 | `Latest ]
   | `CentOS of [ `V6 | `V7 | `V8 | `Latest ]
   | `Debian of [ `V10 | `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
-  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `Latest ]
+  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `V32 | `Latest ]
   | `OracleLinux of [ `V7 | `Latest ]
   | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `Latest ]
   | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `V19_04 | `V19_10 | `V20_04 | `LTS | `Latest ]
@@ -91,12 +91,13 @@ val latest_tag_of_distro : t -> string
   regularly rewritten to point to any new releases of the
   distribution. *)
 
-val base_distro_tag : t -> string * string
-(** [base_distro_tag t] will return a tuple of a Docker Hub
+val base_distro_tag : ?arch:Ocaml_version.arch -> t -> string * string
+(** [base_distro_tag ?arch t] will return a tuple of a Docker Hub
  user/repository and tag for which the base image of a distribution
  can be found (e.g. [opensuse/leap],[15.0] which maps to [opensuse/leap:15.0]
  on the Docker Hub).  This base image is in turn can be used to generate opam
- and other OCaml tool Dockerfiles. *)
+ and other OCaml tool Dockerfiles. [arch] defaults to [x86_64] and can vary
+ the base user/repository since some architecture are built elsewhere. *)
 
 (** {2 CPU architectures} *)